storage: elasticsearch-jest -> elasticsearch-http
Renaming -jest module to -http.
