+++ pom.xml
@@ -210,7 +210,7 @@
 
       <dependency>
         <groupId>${project.groupId}</groupId>
-        <artifactId>zipkin-storage-elasticsearch-jest</artifactId>
+        <artifactId>zipkin-storage-elasticsearch-http</artifactId>
         <version>${project.version}</version>
       </dependency>
 
@@ -222,7 +222,7 @@
 
       <dependency>
         <groupId>${project.groupId}</groupId>
-        <artifactId>zipkin-autoconfigure-storage-elasticsearch-jest</artifactId>
+        <artifactId>zipkin-autoconfigure-storage-elasticsearch-http</artifactId>
         <version>${project.version}</version>
       </dependency>
