+++ zipkin-autoconfigure/storage-elasticsearch-http/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,2 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+zipkin.autoconfigure.storage.elasticsearch.http.ZipkinElasticsearchStorageHttpAutoConfiguration
