--- zipkin-autoconfigure/storage-elasticsearch-jest/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-zipkin.autoconfigure.storage.elasticsearch.jest.ZipkinElasticsearchStorageJestAutoConfiguration
