+++ zipkin-storage/pom.xml
@@ -36,6 +36,6 @@
     <module>cassandra3</module>
     <module>mysql</module>
     <module>elasticsearch</module>
-    <module>elasticsearch-jest</module>
+    <module>elasticsearch-http</module>
   </modules>
 </project>
